replaces DMatrix and Double with FMatrix and Float, tests all test files, HW2, Main

+++ src/extensions/Processing.kt
@@ -2,12 +2,12 @@ package extensions
 
 import processing.core.PApplet
 
-fun PApplet.circleXZ(x: Double, z: Double, radius: Double) {
+fun PApplet.circleXZ(x: Float, z: Float, radius: Float) {
     beginShape()
     val resolution = 20
     for (i in 1..resolution) {
         val theta = 2 * PApplet.PI / (resolution - 1) * i
-        vertex((x + radius * PApplet.cos(theta)).toFloat(), 0f, (z + radius * PApplet.sin(theta)).toFloat())
+        vertex(x + radius * PApplet.cos(theta), 0f, z + radius * PApplet.sin(theta))
     }
     endShape(PApplet.CLOSE)
 }
